Fixed errors in the project.
-For anyone who didn't previosly know: You should never commit your
changes unless u know it doesn't crash and is error free. If your
project has a red box with an X in it that means you have an error
somewhere and it will most likely crash when you load the mod.

-Moved ItemLightCoffee to com.team.kalstuff.item.

-Added needed imports in ItemLightCoffee.

-Added the ItemLightCoffee variable in StartupCommon and registered it
with GameRegistry.

-Added the ModelMesher for ItemLightCoffee. Now it renders right.

Pie, you might want to look at these changes so you can know all the
steps to adding a new item.